+++ cgroup/v1/blkio/mount.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+mkdir -p blkio.mnt
+sudo mount -t cgroup -o blkio none blkio.mnt