+++ cgroup/v1/blkio/test.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-sudo mount -t cgroup -o blkio none /sys/fs/cgroup/blkio
+./mount.sh
 
 # Specify a bandwidth rate on particular device for root group. The format for
 # policy is
@@ -10,6 +10,6 @@ sudo mount -t cgroup -o blkio none /sys/fs/cgroup/blkio
 #
 # This will put a limit of 1MB/second on reads happening for root group on
 # device having major/minor number 8:16.
-echo "8:16  1048576" | sudo tee /sys/fs/cgroup/blkio/blkio.throttle.read_bps_device
+echo "8:16  1048576" | sudo tee blkio.mnt/blkio.throttle.read_bps_device
 
 # Then, test with dd.